feat: move new upgrades scripts in accessors subfolder
